# tests: speedup readhistory reduce test