For example on my machine:

| test duration (ms) |  master  |             PR |
| :-                 | -:       | -:             |
| regular run        |  2645.94 | 106.83 (÷24.8) |
| coverage run       | 26133.27 | 641.59 (÷40.7) |

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/koreader/koreader/12785)
<!-- Reviewable:end -->